feat:enhance org, support outputing c++ asm as results
I mimic the codes in ob-C.el which defines the execution of C/C++ src
block in org mode and write a new babel language called cpp-asm. The
codes are stored in lisp/org-custom-babel.el. It basically does the
thing like org-babel-execute:C++. But instead of emitting an
executable and execute it, org-babel-execue:cpp-asm emits a assembly
file and read it into results. The new language is bound to c++-mode.